PR_1 Обработаны замечания
